requirements.txt
ipython==8.12.3
matplotlib==3.8.4
numpy==1.26.4
opencv_python==4.8.1.78
Pillow==10.1.0
Pillow==10.3.0
scikit_learn==1.4.2
tensorflow==2.16.1
